backend/pyproject.toml
[tool.poetry]
name = "splitsphere"
version = "0.1.0"
description = ""
authors = ["Vivek M. Jude <vivekmjude@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.11"
uvicorn = { extras = ["standard"], version = "^0.34.0" }
sqlalchemy = "^2.0.38"
pydantic = ">=2.0"
alembic = "^1.15.1"
psycopg2-binary = "^2.9.10"
python-multipart = "^0.0.20"
python-jose = { extras = ["cryptography"], version = "^3.4.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-dotenv = "^1.0.1"
httpx = "^0.28.1"
supabase = "^2.13.0"
pydantic-settings = "^2.8.1"
isort = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.25.3"
ruff = "^0.9.10"
black = "^25.1.0"
mypy = "^1.15.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
